Time: 11 ms (65.05%), Space: 17.3 MB (54.77%) - LeetHub

+++ 662-maximum-width-of-binary-tree/662-maximum-width-of-binary-tree.cpp
@@ -5,7 +5,7 @@ public:
             return 0;
         
         int res = 1;
-        queue<pair<TreeNode*, int>> q;
+        queue<pair<TreeNode*, unsigned int>> q;
         
         // I am using intialising list
         q.push({root, 0});      // also can use make_pair
@@ -25,17 +25,17 @@ public:
                 // we will use it while inserting it children
                 // left child will be 2 * idx + 1;
                 // right chils will be 2 * idx + 2;
-                int idx = p.second - start;
+                unsigned int idx = p.second - start;
                 
                 q.pop();
                 
                 // if  left child exist
                 if(p.first->left != NULL)
-                    q.push({p.first->left, (long long)2 * idx + 1});
+                    q.push({p.first->left, 2 * idx + 1});
                 
                 // if right child exist
                 if(p.first->right != NULL)
-                    q.push({p.first->right, (long long) 2 * idx + 2});
+                    q.push({p.first->right,  2 * idx + 2});
             }
         }
